# src/docs/openapi.yaml
openapi: "3.0.0"
info:
  version: 6.0.0
  title: Jeera Project Manager
  license:
    name: MIT
  contact:
    name: Antoine Cheron
    url: https://antoinecheron.github.io
    email: antoine.cheron@irisa.fr
servers:
  - url: http://localhost:3000/
paths:
...
  /project/{projectId}/task/{taskId}:
    parameters: 
      - $ref: '#/components/parameters/projectId'
      - $ref: '#/components/parameters/taskId'
    get:
      summary: Get the detail of a task
      operationId: getProjectTaskDetails
      x-@id: vnd_jeera:getTaskDetails
      tags:
        - tasks
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
          links:
            update:
              operationId: updateProjectTask
              x-@relation: vnd_jeera:rel/modify
              parameters:
                taskId: '$response.body#/id'
                projectId: '$response.body#/parentProjectId'
                id: '$response.body#/id'
                title: '$response.body#/title'
                description: '$response.body#/description'
                assignee: '$response.body#/assignee'
                points: '$response.body#/points'
                status: '$response.body#/status'
                tags: '$response.body#/tags'
                priority: '$response.body#/priority'
            delete:
              operationId: deleteProjectTask
              x-@relation: vnd_jeera:rel/delete
              parameters:
                taskId: '$response.body#/id'
                projectId: '$response.body#/parentProjectId'
            moveToQA:
              operationId: toQATask
              x-@relation: vnd_jeera:rel/toQATask
              parameters:
                taskId: '$response.body#/id'
                projectId: '$response.body#/parentProjectId'
            complete:
              operationId: toDoneTask
              x-@relation: vnd_jeera:rel/completeTask
              parameters:
                taskId: '$response.body#/id'
                projectId: '$response.body#/parentProjectId'
            analytics:
              operationId: getAnalytics
              x-@relation: vnd_jeera:rel/analytics
              parameters:
                resourceId: '$response.body#/id'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      summary: Delete a task (only if archived)
      operationId: deleteProjectTask
      x-@id: vnd_jeera:deleteProjectTask
      tags:
        - tasks
      responses:
        '204':
          description: Success
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security: 
        - bearerToken: []
    put:
      summary: Update some fields of the task
      operationId: updateProjectTask
      x-@id: vnd_jeera:updateProjectTask
      tags:
        - tasks
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/TechnicalStoryUpdateRequest'
                - $ref: '#/components/schemas/UserStoryUpdateRequest'
      responses:
        '200':
          description: Update successfully done
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
          links:
            update:
              operationId: updateProjectTask
              x-@relation: vnd_jeera:rel/modify
              parameters:
                taskId: '$response.body#/id'
                projectId: '$response.body#/parentProjectId'
                id: '$response.body#/id'
                title: '$response.body#/title'
                description: '$response.body#/description'
                assignee: '$response.body#/assignee'
                points: '$response.body#/points'
                status: '$response.body#/status'
                tags: '$response.body#/tags'
                priority: '$response.body#/priority'
            delete:
              operationId: deleteProjectTask
              x-@relation: vnd_jeera:rel/delete
              parameters:
                taskId: '$response.body#/id'
                projectId: '$response.body#/projectId'
            moveToQA:
              operationId: toQATask
              x-@relation: vnd_jeera:rel/toQATask
              parameters:
                taskId: '$response.body#/id'
                projectId: '$response.body#/projectId'
            complete:
              operationId: toDoneTask
              x-@relation: vnd_jeera:rel/completeTask
              parameters:
                taskId: '$response.body#/id'
                projectId: '$response.body#/projectId'
            analytics:
              operationId: getAnalytics
              x-@relation: vnd_jeera:rel/analytics
              parameters:
                resourceId: '$response.body#/id'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security: 
        - bearerToken: []
...
  /project/{projectId}/tasks:
    parameters: 
        - $ref: '#/components/parameters/projectId'
    get:
      summary: List user tasks for a given project
      operationId: listProjectTasks
      x-@id: vnd_jeera:listTasks
      tags:
        - tasks
      parameters:
        - name: createdBefore
          in: query
          x-@id: vnd_jeera:createdBefore
          description: Minimum date of tasks to list
          required: false
          schema:
            type: string
            format: date
        - name: offset
          in: query
          x-@id: hydra:offset
          description: The number of items to skip before starting to collect the result set
          required: false
          schema:
            type: integer
            format: int8
            minimum: 0
            default: 0
        - name: limit
          in: query
          x-@id: hydra:limit
          description: Number of elements to get
          required: false
          schema:
            type: integer
            format: int8
            minimum: 1
            maximum: 40
            default: 10
      responses:
        '200':
          description: Success - list of tasks
          headers:
            X-Next:
              description: Link to next page
              x-@id: hydra:next
              required: false
              schema:
                type: string
                format: url
            X-Last:
              description: Link to last page
              x-@id: hydra:last
              required: false
              schema:
                type: string
                format: url
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TasksList'
          links:
            createTechnicalStory:
              operationId: createTechnicalStory
              x-@relation: vnd_jeera:rel/create
              parameters:
                projectId: '$request.path.projectId'
            createUserStory:
              operationId: createUserStory
              x-@relation: vnd_jeera:rel/create
              parameters:
                projectId: '$request.path.projectId'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security: 
        - bearerToken: []
  /project/{projectId}/tasks/technicalStory:
    parameters: 
        - $ref: '#/components/parameters/projectId'
    post:
      summary: Creates a technical story
      operationId: createTechnicalStory
      x-@id: vnd_jeera:createTechnicalStory
      tags:
        - tasks
      requestBody:
        description: Technical story to create
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TechnicalStoryCreationRequest'
      responses:
        '201':
          description: Created technical story
          headers:
            Location:
              description: 'URL to the resource created'
              schema:
                type: string
                format: url
                x-@type: '@id'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TechnicalStory'
          links:
            update:
              operationId: updateProjectTask
              x-@relation: vnd_jeera:rel/modify
              parameters:
                taskId: '$response.body#/id'
                projectId: '$response.body#/parentProjectId'
                id: '$response.body#/id'
                title: '$response.body#/title'
                description: '$response.body#/description'
                assignee: '$response.body#/assignee'
                points: '$response.body#/points'
                status: '$response.body#/status'
                tags: '$response.body#/tags'
                priority: '$response.body#/priority'
            delete:
              operationId: deleteProjectTask
              x-@relation: vnd_jeera:rel/delete
              parameters:
                taskId: '$response.body#/id'
                projectId: '$response.body#/parentProjectId'
            moveToQA:
              operationId: toQATask
              x-@relation: vnd_jeera:rel/toQATask
              parameters:
                taskId: '$response.body#/id'
                projectId: '$response.body#/parentProjectId'
            complete:
              operationId: toDoneTask
              x-@relation: vnd_jeera:rel/completeTask
              parameters:
                taskId: '$response.body#/id'
                projectId: '$response.body#/parentProjectId'
            analytics:
              operationId: getAnalytics
              x-@relation: vnd_jeera:rel/analytics
              parameters:
                resourceId: '$response.body#/id'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security: 
        - bearerToken: []
  /project/{projectId}/tasks/userStory:
    parameters: 
      - $ref: '#/components/parameters/projectId'
    post:
      summary: Creates a user story
      operationId: createUserStory
      x-@id: vnd_jeera:createUserStory
      tags:
        - tasks
      requestBody:
        description: User story to create
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserStoryCreationRequest'
      responses:
        '201':
          description: Created user story
          headers:
            Location:
              description: 'URL to the resource created'
              schema:
                type: string
                format: url
                x-@type: '@id'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserStory'
          links:
            update:
              operationId: updateProjectTask
              x-@relation: vnd_jeera:rel/modify
              parameters:
                taskId: '$response.body#/id'
                projectId: '$response.body#/parentProjectId'
                id: '$response.body#/id'
                title: '$response.body#/title'
                description: '$response.body#/description'
                assignee: '$response.body#/assignee'
                points: '$response.body#/points'
                status: '$response.body#/status'
                tags: '$response.body#/tags'
                priority: '$response.body#/priority'
            delete:
              operationId: deleteProjectTask
              x-@relation: vnd_jeera:rel/delete
              parameters:
                taskId: '$response.body#/id'
                projectId: '$response.body#/parentProjectId'
            moveToQA:
              operationId: toQATask
              x-@relation: vnd_jeera:rel/toQATask
              parameters:
                taskId: '$response.body#/id'
                projectId: '$response.body#/parentProjectId'
            complete:
              operationId: toDoneTask
              x-@relation: vnd_jeera:rel/completeTask
              parameters:
                taskId: '$response.body#/id'
                projectId: '$response.body#/parentProjectId'
            analytics:
              operationId: getAnalytics
              x-@relation: vnd_jeera:rel/analytics
              parameters:
                resourceId: '$response.body#/id'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security: 
        - bearerToken: []
  /project/{projectId}/task/{taskId}/toQa:
    parameters: 
      - $ref: '#/components/parameters/projectId'
      - $ref: '#/components/parameters/taskId'
    put:
      description: Place the task in QA status
      operationId: toQATask
      x-@id: vnd_jeera:toQATask
      tags:
        - tasks
      responses:
        '204':
          description: Success
          links:
            modify:
              operationId: updateProjectTask
              x-@relation: vnd_jeera:rel/modify
              parameters:
                taskId: '$response.body#/id'
                projectId: '$response.body#/parentProjectId'
                id: '$response.body#/id'
                title: '$response.body#/title'
                description: '$response.body#/description'
                assignee: '$response.body#/assignee'
                points: '$response.body#/points'
                status: '$response.body#/status'
                tags: '$response.body#/tags'
                priority: '$response.body#/priority'
            complete:
              operationId: toDoneTask
              x-@relation: vnd_jeera:rel/completeTask
              parameters:
                taskId: '$response.body#/id'
                projectId: '$response.body#/parentProjectId'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security: 
        - bearerToken: []
  /project/{projectId}/task/{taskId}/complete:
    parameters: 
      - $ref: '#/components/parameters/projectId'
      - $ref: '#/components/parameters/taskId'
    put:
      description: Place the task in done status
      operationId: toDoneTask
      x-@id: vnd_jeera:completeTask
      tags:
        - tasks
      responses:
        '204':
          description: Success
          links:
            modify:
              operationId: updateProjectTask
              x-@relation: vnd_jeera:rel/modify
              parameters:
                taskId: '$response.body#/id'
                projectId: '$response.body#/parentProjectId'
                id: '$response.body#/id'
                title: '$response.body#/title'
                description: '$response.body#/description'
                assignee: '$response.body#/assignee'
                points: '$response.body#/points'
                status: '$response.body#/status'
                tags: '$response.body#/tags'
                priority: '$response.body#/priority'
            moveToQA:
              operationId: toQATask
              x-@relation: vnd_jeera:rel/toQATask
              parameters:
                taskId: '$response.body#/id'
                projectId: '$response.body#/parentProjectId'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security: 
        - bearerToken: []
...
components:
  securitySchemes:
    bearerToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  parameters:
...
    taskId:
      name: taskId
      in: path
      x-@id: vnd_jeera:taskId
      description: Id of the task to retrieve details for
      required: true
      schema:
        type: string
        format: url
        x-@type: '@id'
    projectId:
      name: projectId
      in: path
      x-@id: vnd_jeera:projectId
      description: Id of the project to retrieve details for
      required: true
      schema:
        type: string
        format: url
        x-@type: '@id'
...
  schemas:
...
    TasksList:
      type: object
      properties:
        tasks:
          type: array
          items:
            $ref: '#/components/schemas/Task'
    TechnicalStoryCreationRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 4
          maxLength: 80
        description:
          type: string
          minLength: 0
          maxLength: 4000
        assignee:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/TaskStatusFreeToMove'
        tags:
          type: array
          items:
            type: string
          maxItems: 10
        priority:
          $ref: '#/components/schemas/Priority'
      required: [ title, assignee ]
    TechnicalStoryUpdateRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 4
          maxLength: 80
        description:
          type: string
          minLength: 0
          maxLength: 4000
        assignee:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/TaskStatusFreeToMove'
        tags:
          type: array
          items:
            type: string
          maxItems: 10
        priority:
          $ref: '#/components/schemas/Priority'
    TechnicalStory:
      type: object
      properties:
        id:
          type: string
          format: uuid
          x-@id: vnd_jeera:taskId
          x-@type: 'mo:uuid'
        title:
          type: string
        parentProjectId:
          type: string
          format: uuid
        description:
          type: string
        assignee:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/TaskStatus'
        tags:
          type: array
          items:
            type: string
        priority:
          $ref: '#/components/schemas/Priority'
      required: [ id, title, parentProjectId, description, assignee, status ]
    UserStoryCreationRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 4
          maxLength: 80
        description:
          type: string
          minLength: 0
          maxLength: 4000
        assignee:
          type: string
          format: uuid
        points:
          type: number
          minimum: 0
          maximum: 120
        status:
          $ref: '#/components/schemas/TaskStatusFreeToMove'
        tags:
          type: array
          items:
            type: string
          maxItems: 10
        priority:
          $ref: '#/components/schemas/Priority'
      required: [ title, assignee, points ]
    UserStoryUpdateRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 4
          maxLength: 80
        description:
          type: string
          minLength: 0
          maxLength: 4000
        assignee:
          type: string
          format: uuid
        points:
          type: number
          minimum: 0
          maximum: 120
        status:
          $ref: '#/components/schemas/TaskStatusFreeToMove'
        tags:
          type: array
          items:
            type: string
          maxItems: 10
        priority:
          $ref: '#/components/schemas/Priority'
    UserStory:
      type: object
      properties:
        id:
          type: string
          format: uuid
          x-@id: vnd_jeera:taskId
          x-@type: 'mo:uuid'
        title:
          type: string
        points:
          type: number
        parentProjectId:
          type: string
          format: uuid
        description:
          type: string
        assignee:
          type: string
          format: uuid
        creationDate:
          type: string
          format: date
        lastUpdate:
          type: string
          format: date
        status:
          $ref: '#/components/schemas/TaskStatus'
        tags:
          type: array
          items:
            type: string
        priority:
          $ref: '#/components/schemas/Priority'
        updatesCount:
          type: integer
      required: [ id, title, points, parentProjectId, description, assignee, lastUpdate, status, updatesCount ]
    Task:
      oneOf:
        - $ref: '#/components/schemas/TechnicalStory'
        - $ref: '#/components/schemas/UserStory'
    TaskStatus:
      type: string
      enum: [ 'todo', 'in progress', 'review', 'QA', 'done' ]
    TaskStatusFreeToMove:
      type: string
      enum: [ 'todo', 'in progress', 'review' ]
    Priority:
      type: string
      enum:
        - blocking
        - critical
        - important
        - high
        - medium
        - low
        - simple
...
    Error:
      type: object
      properties:
        errorCode:
          type: string
          x-@id: schema:identifier
...